requirements.txt
Flask
transformers
torch
scikit-learn
python-dotenv
